Resume-screening-ai-main/backend/requirements.txt
fastapi
uvicorn
groq
pdfplumber
python-docx
python-multipart
python-dotenv